change a ref to links and changed names from classes to courses also set up components to accept data for courses
