+++ src/components/pages/InstructorHome/InstructorSidebar.js
@@ -34,19 +34,19 @@ function InstructorSidebar() {
       <div className="logo" />
       <Menu theme="light" defaultSelectedKeys={['1']} mode="inline">
         <Menu.Item key="2" icon={<HomeOutlined />}>
-          <a href="/instructor">Overview</a>
+          <Link to="/instructor">Overview</Link>
         </Menu.Item>
 
         <Menu.Item key="3" icon={<ReadOutlined />}>
-          <a href="/instructor-all-classes">Courses</a>
+          <Link to="/instructor-all-classes">Courses</Link>
         </Menu.Item>
 
         <Menu.Item key="4" icon={<ReadOutlined />}>
-          <a href="/classroom">Course</a>
+          <Link to="/classroom">Course</Link>
         </Menu.Item>
 
         <Menu.Item key="5" icon={<PlusOutlined />}>
-          <a href="/messages">Messages</a>
+          <Link to="/messages">Messages</Link>
         </Menu.Item>
 
         <Menu.Item
